+++ extensions/apm/pom.xml
@@ -4,17 +4,17 @@
 
     <parent>
         <groupId>io.github.onecx.quarkus</groupId>
-        <artifactId>onecx-quarkus</artifactId>
+        <artifactId>onecx-quarkus-extensions-parent</artifactId>
         <version>999-SNAPSHOT</version>
         <relativePath>../pom.xml</relativePath>
     </parent>
 
-    <artifactId>onecx-quarkus-extensions-parent</artifactId>
+    <artifactId>onecx-apm-parent</artifactId>
     <packaging>pom</packaging>
 
     <modules>
-        <module>parameters</module>
-        <module>apm</module>
+        <module>deployment</module>
+        <module>runtime</module>
     </modules>
 
 </project>
