+++ extensions/apm/runtime/src/main/java/io/github/onecx/quarkus/apm/client/PermissionDTOV3.java
@@ -1,0 +1,9 @@
+package io.github.onecx.quarkus.apm.client;
+
+import java.util.HashMap;
+import java.util.List;
+
+//
+public class PermissionDTOV3 extends HashMap<String, List<String>> {
+
+}
