+++ extensions/parameters/deployment/src/main/java/io/github/onecx/quarkus/it/parameter/ParametersBuildTimeConfig.java
@@ -1,0 +1,18 @@
+package io.github.onecx.quarkus.it.parameter;
+
+import io.github.onecx.quarkus.it.parameter.devservices.DevServicesConfig;
+import io.quarkus.runtime.annotations.ConfigItem;
+import io.quarkus.runtime.annotations.ConfigRoot;
+
+@ConfigRoot(name = "parameters", prefix = "onecx")
+public class ParametersBuildTimeConfig {
+
+    /**
+     * Configuration for DevServices
+     * <p>
+     * DevServices allows Quarkus to automatically start MockServer in dev and test mode.
+     */
+    @ConfigItem(name = "devservices")
+    public DevServicesConfig devServices;
+
+}
