+++ it/apm/src/main/resources/application.properties
@@ -1,0 +1,23 @@
+quarkus.banner.enabled=false
+
+# TEST
+%test.quarkus.keycloak.devservices.roles.alice=role-admin
+%test.quarkus.keycloak.devservices.roles.bob=role-user
+%test.quarkus.mockserver.devservices.config-class-path=true
+%test.quarkus.mockserver.devservices.config-file=/mockserver.properties
+%test.quarkus.mockserver.devservices.config-dir=/mockserver
+%test.quarkus.mockserver.devservices.log=true
+%test.quarkus.rest-client.onecx-apm.url=${quarkus.mockserver.endpoint}
+%test.onecx.apm.log=true
+%test.onecx.apm.application-id=app-id
+%test.org.eclipse.microprofile.rest.client.propagateHeaders=Authorization,apm-principal-token
+
+# DEV
+%dev.quarkus.keycloak.devservices.roles.alice=role-admin
+%dev.quarkus.keycloak.devservices.roles.bob=role-user
+%dev.quarkus.mockserver.devservices.config-file=src/test/resources/mockserver.properties
+%dev.quarkus.mockserver.devservices.config-dir=src/test/resources/mockserver
+%dev.quarkus.mockserver.devservices.log=true
+%dev..quarkus.rest-client.onecx-apm.url=${quarkus.mockserver.endpoint}
+%dev.onecx.apm.log=true
+%dev.onecx.apm.application-id=app-id
