+++ it/apm/src/test/resources/mockserver.properties
@@ -1,0 +1,16 @@
+mockserver.initializationJsonPath=/mockserver/*.json
+# watch changes in the file
+mockserver.watchInitializationJson=true
+
+
+# Certificate Generation
+# dynamically generated CA key pair (if they don't already exist in specified directory)
+mockserver.dynamicallyCreateCertificateAuthorityCertificate=true
+# save dynamically generated CA key pair in working directory
+mockserver.directoryToSaveDynamicSSLCertificate=.
+# certificate domain name (default "localhost")
+mockserver.sslCertificateDomainName=localhost
+# comma separated list of ip addresses for Subject Alternative Name domain names (default empty list)
+mockserver.sslSubjectAlternativeNameDomains=www.example.com,www.another.com
+# comma separated list of ip addresses for Subject Alternative Name ips (default empty list)
+mockserver.sslSubjectAlternativeNameIps=127.0.0.1
