+++ it/parameters/src/test/java/io/github/onecx/quarkus/it/parameter/ParametersIT.java
@@ -1,0 +1,7 @@
+package io.github.onecx.quarkus.it.parameter;
+
+import io.quarkus.test.junit.QuarkusIntegrationTest;
+
+@QuarkusIntegrationTest
+public class ParametersIT extends ParametersTest {
+}
