--- it/parameters/src/test/java/io/github/onecx/quarkus/parameter/ParametersIT.java
@@ -1,7 +1,0 @@
-package io.github.onecx.quarkus.parameter;
-
-import io.quarkus.test.junit.QuarkusIntegrationTest;
-
-@QuarkusIntegrationTest
-public class ParametersIT extends ParametersTest {
-}
